pyproject.toml
[project]
name = "upgraded-couscous"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-bootstrap>=3.3.7.1",
    "flask-sqlalchemy>=3.1.1",
    "newsapi-python>=0.2.7",
    "python-dotenv>=1.2.2",
]
